Update folder name and refactor components with React.FC
