+++ Web/starter_project_vanguard/components/success-story/Partners.tsx
@@ -1,4 +1,6 @@
-const Partners = () => {
+import React from "react"
+
+const Partners: React.FC = () => {
   return (
     <div className="max-w-screen-lg mt-20">
       <h1 className="font-DMSans text-5xl font-semibold mb-16 text-center">
@@ -6,29 +8,29 @@ const Partners = () => {
       </h1>
       <div className="flex flex-col sm:flex-row  gap-1">
         <img
-          src="/images/story-img/Google.png"
+          src="/images/success-story/google.png"
           alt="Google logo"
           className="w-full max-w-screen-lg sm:max-w-none"
         />
         <img
-          src="/images/story-img/Palantir.png"
+          src="/images/success-story/palantir.png"
           alt="Palantir logo"
           className="w-full max-w-screen-lg sm:max-w-none"
         />
         <img
-          src="/images/story-img/InstDeep.png"
+          src="/images/success-story/inst-deep.png"
           alt="InstaDeep logo"
           className="w-full max-w-screen-lg sm:max-w-none"
         />
         <img
-          src="/images/story-img/meta.png"
+          src="/images/success-story/meta.png"
           alt="Meta logo"
           className="w-full max-w-screen-lg sm:max-w-none"
         />
       </div>
       <div className="flex gap-1 flex-col sm:flex-row justify-between mx-0 md:mx-40">
-        <img src="/images/story-img/databricks.png" alt="Databricks log" />
-        <img src="/images/story-img/linkedin.png" alt="LinkedIn logo" />
+        <img src="/images/success-story/databricks.png" alt="Databricks logo" />
+        <img src="/images/success-story/linkedin.png" alt="LinkedIn logo" />
       </div>
     </div>
   )
