--- Web/starter_project_vanguard/pages/sucess-story.tsx
@@ -1,81 +1,0 @@
-import Partners from '@/components/story/Partners'
-import Text from '@/components/story/Text'
-import { mockData } from '../utils/data/mockdata'
-import Head from 'next/head'
-import SucessImage from '@/components/story/story-img'
-
-export const metadata = {
-  title: 'Sucess Story',
-  description: 'Generated by create next app',
-}
-
-export default function SucessStory() {
-  return (
-    <>
-       <Head>
-          <link
-            href="https://fonts.googleapis.com/css2?family=Neucha&display=swap"
-            rel="stylesheet"
-          />
-          <link
-          href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700;800;900&family=Roboto&display=swap"
-          rel="stylesheet"
-        />
-        <title> Success Stories</title>
-        </Head>
-    <main>
-      <div className="flex justify-center mt-4">
-        <h1 className="font-poppins text-5xl font-semibold">Impact Stories</h1>
-      </div>
-      <div className="flex justify-center mx-2 mt-2">
-        <div className="flex justify-center max-w-xl">
-          <p className="text-center font-poppins text-2xl">
-            Behind every success is a story. Learn about the stories of A2SVians
-          </p>
-        </div>
-      </div>
-      <div className="flex justify-center">
-        <hr className="border-2 w-10 rounded-sm border-blue-700" />
-      </div>
-      <div className="flex justify-center">
-        <div className="max-w-5xl">
-          <div>
-            {mockData.map((mock_data, index) => (
-              <div
-                key={index}
-                className={`flex flex-col lg:flex-row justify-center ml-2 mt-20 ${
-                  index % 2 === 1 ? 'lg:flex-row-reverse' : ''
-                }`}
-              >
-                <div className="  lg:block">
-                  <SucessImage
-                    name={mock_data.name}
-                    image={mock_data.image}
-                    category={mock_data.category}
-                    location={mock_data.location}
-                  />
-                </div>
-                <div className="mt-0">
-                  <Text
-                    heading1={mock_data.heading1}
-                    paragraph={mock_data.paragraph1}
-                  />
-                  <Text
-                    heading1={mock_data.heading2}
-                    paragraph={mock_data.paragraph2}
-                  />
-                  <Text
-                    heading1={mock_data.heading3}
-                    paragraph={mock_data.paragraph3}
-                  />
-                </div>
-              </div>
-            ))}
-          </div>
-          <Partners />
-        </div>
-      </div>
-    </main>
-    </>
-  )
-}
